#117 - refactoring
